+++ src/wavekit/scope.py
@@ -253,7 +253,7 @@ def match_signals(
             for k, p in pattern_list[0].items():
                 if p[0] == '@':
                     name_regex, range_suffix = split_by_range_expr(p[1:])
-                    if match := re.fullmatch(name_regex, sig.name):
+                    if match := re.fullmatch(name_regex, sig_bare):
                         assert len(k) == 0
                         key = (match.groups(),)
                         if len(pattern_list) == 1: